refactor: move account branch component to distinct file
